+++ NEWS.md
@@ -1,3 +1,9 @@
+# poissoned 0.1.1
+
+* Added arguments `xinit`, `yinit` to allow the user to specify the initial seed point
+* Added argument `keep_idx` to allow the user to keep the order in which the points
+were generated.
+
 # poissoned 0.1.0
 
 * Initial release